backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0

SQLAlchemy[asyncio]>=2.0,<3.0
asyncpg>=0.29,<1.0

pydantic>=2.5,<3.0
email-validator>=2.1,<3.0
python-multipart>=0.0.9,<1.0

passlib[bcrypt]>=1.7.4,<2.0
PyJWT>=2.8,<3.0

minio>=7.2,<8.0

pytesseract>=0.3.10,<1.0
Pillow>=10.2,<11.0

reportlab>=4.0,<5.0

pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
httpx>=0.27,<1.0

